Converted HuobiExchangeAdapter to take all config from exchange.xml. Currently undergoing live testing with the exchange.
Also added UnknownHostException to Abstract adapter catch list - EC2 started barfing with these for both BTC-e and GDAX (at same time) on 14 July 2016.

The following adapters will not work (until they are converted) with the current SNAPSHOT version of the build: Bitfinex.

Stable versions of the bot can always be found on the Release page.